requirements.txt
aiogram==2.23.1
APScheduler==3.10.1
sqlite==0.6.0
SQLAlchemy==1.4.41
python-dotenv==1.0.0
pytz==2023.3
xlsxwriter==3.0.3